feat: make sure chadrc.lua is never overwritten during install

+++ install.zsh
@@ -11,8 +11,11 @@ NVIM_HOME="$(realpath ~)/.config/nvim"
 ROOT=$(realpath ../)
 CONFIG=$(realpath ./config)
 
-# Copy chadrc.lua.template to chadrc.lua
-cp ./config/chadrc.lua.template ./config/chadrc.lua
+# Copy chadrc.lua.template to chadrc.lua (if it doesn't already exist)
+if [ ! -e "config/chadrc.lua" ]
+then
+  cp ./config/chadrc.lua.template ./config/chadrc.lua
+fi
 
 # Clone NvChad along side this folder
 if [ ! -d "../NvChad" ]
